requirements.txt
Django>=3.2,<4.0
mysqlclient>=2.0.3
djangorestframework>=3.12.4
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=3.7.0
